third_party/WebKit/LayoutTests/platform/win/virtual/mojo-loading/http/tests/security/contentSecurityPolicy/require-sri-for/require-sri-for-style-blocked-expected.txt

CONSOLE ERROR: line 15: Refused to load the stylesheet 'http://127.0.0.1:8000/security/contentSecurityPolicy/resources/style-set-red.css' because 'require-sri-for' directive requires integrity attribute be present for all stylesheets.
This is a testharness.js-based test.
PASS Stylesheets without integrity generate reports. 
PASS Stylesheets without integrity do not load. 
Harness: the test ran to completion.
